+++ src/node-locations.cpp
@@ -67,9 +67,9 @@ osmium::Location node_locations_t::get(osmid_t id) const
     for (std::size_t n = 0; n < block_size && begin != end; ++n) {
         auto const bid = did.update(
             static_cast<int64_t>(protozero::decode_varint(&begin, end)));
-        auto const x = static_cast<int32_t>(dx.update(
+        auto const x = static_cast<int64_t>(dx.update(
             protozero::decode_zigzag64(protozero::decode_varint(&begin, end))));
-        auto const y = static_cast<int32_t>(dy.update(
+        auto const y = static_cast<int64_t>(dy.update(
             protozero::decode_zigzag64(protozero::decode_varint(&begin, end))));
         if (bid == id) {
             return osmium::Location{x, y};
